More unit tests.
