+++ tests/array_test.go
@@ -71,21 +71,3 @@ func TestContractString(t *testing.T) {
 		t.Error(err)
 	}
 }
-
-func TestCumulativeU256(t *testing.T) {
-	currentPath, _ := os.Getwd()
-	targetPath := path.Join(path.Dir(filepath.Dir(currentPath)), "concurrentlib/lib/")
-	err, _, _ := DeployThenInvoke(targetPath, "commutative/u256Cum_test.sol", "0.8.19", "CumulativeU256Test", "call()", []byte{}, false)
-	if err != nil {
-		t.Error(err)
-	}
-}
-
-func TestCumulativeU256Counter(t *testing.T) {
-	currentPath, _ := os.Getwd()
-	targetPath := path.Join(path.Dir(filepath.Dir(currentPath)), "concurrentlib/lib/")
-	err, _, _ := DeployThenInvoke(targetPath, "commutative/u256Cum_test.sol", "0.8.19", "VisitCounter", "call()", []byte{}, false)
-	if err != nil {
-		t.Error(err)
-	}
-}
